requirements.txt
asgiref==3.6.0
backports.zoneinfo;python_version<"3.9"
Django==4.2
sqlparse==0.4.3
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
rest-framework-simplejwt==0.0.2
PyJWT==2.6.0